# BookWormApi/swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"Book Worm API"},"paths":{"/api/Text/Sort/{sortType}":{"post":{"tags":["TextParsing"],"summary":"Takes a text and sorts it according to the sort type given.\r\nIf the text contains only one sentence it will sort the words in \r\nthe sentence.\r\nOtherwise it will sort the sentences.","operationId":"Post","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"input","in":"body","description":"BookWormApi.Models.TextParsingInputThe input text to sort","required":false,"schema":{"$ref":"#/definitions/TextParsingInput"}},{"name":"sortType","in":"path","description":"BookWormApi.SortTypeThe type of sorting to apply","required":true,"type":"string","enum":["alphabetical","reverse","length"]}],"responses":{"200":{"description":"Success","schema":{"type":"string"}},"500":{"description":"Server Error","schema":{"type":"object"}}}}},"/api/Text/Statistics":{"post":{"tags":["TextParsing"],"summary":"Calculates some basic statistics about the given text.","operationId":"Post","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"input","in":"body","description":"BookWormApi.Models.TextParsingInputThe input text to analyze","required":false,"schema":{"$ref":"#/definitions/TextParsingInput"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/TextStatistics"}},"500":{"description":"Server Error","schema":{"type":"object"}}}}}},"definitions":{"TextParsingInput":{"description":"Data contract for sending requests to the application","type":"object","properties":{"text":{"description":"The text to analyze","type":"string"}}},"TextStatistics":{"description":"Data contract for returning statistics from a text","type":"object","properties":{"wordCount":{"format":"int32","description":"The number of words in the incoming text","type":"integer","readOnly":true},"spaceCount":{"format":"int32","description":"The number of spaces in the incoming text","type":"integer","readOnly":true},"letterCount":{"format":"int32","description":"The number of letters in the incoming text","type":"integer","readOnly":true},"hyphenCount":{"format":"int32","description":"The number of hyphens in the incoming text","type":"integer","readOnly":true},"sentenceCount":{"format":"int32","description":"The number of sentences in the incoming text","type":"integer","readOnly":true}}}}}